requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
mediapipe>=0.10.0
Pillow>=9.0.0
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
